Multiple fixes in payment lifecycle (#693)
* permissive codec for failure messages

Some implementations were including/ommitting the message type when
including a `channel_update` in failure messages.

We now use a codec that supports both versions for decoding, and
will encode *with* the message type.

* make router handle updates from failure messages

This is a regression caused by 9f708acf04f11551ec6a6c9147f1c9fcfba737e7,
which made the `Peer` class encapsulates network announcements inside
`PeerRoutingMessage`, in order to preserve the origin of the messages.

But when we get channel updates as part of failure messages when making
payments, they weren't encapsulated and were ignored by the router.

Re-enabled a non-regression test in `IntegrationSpec` which will prevent
this from happening in the future.

* improve handling of `Update` payment failures

Exclude nodes that send us multiple `Update` failures for the same
payment (they may be bad actors)